cleanup + comments
