+++ lib/Core/filterOutUndefined.ts
@@ -1,7 +1,5 @@
 import { NotUndefined } from "./TypeModifiers";
 
-export default function filterOutUndefined<T>(
-  array: Array<T | undefined>
-): Array<NotUndefined<T>> {
+export default function filterOutUndefined<T>(array: Array<T | undefined>) {
   return array.filter(x => x !== undefined) as Array<NotUndefined<T>>;
 }
