+++ lib/Models/Definition/Model.ts
@@ -107,8 +107,7 @@ export interface ModelInterface<T extends ModelTraits> {
    * @param stratumId The ID of the stratum in which to add the object.
    * @param trait The name of the {@link objectArrayTrait} property.
    * @param objectId The ID of the new object.
-   * @returns The new object, or undefined if the object still does not exist
-   *          because a stratum above the specified one has removed it.
+   * @returns The new object, or undefined if the object still does not exist because a stratum above the specified one has removed it.
    */
   addObject<Key extends keyof ArrayElementTypes<T>>(
     stratumId: string,
@@ -116,6 +115,17 @@ export interface ModelInterface<T extends ModelTraits> {
     objectId: string
   ): Model<ArrayElementTypes<T>[Key]> | undefined;
 
+  /**
+   * Pushes a new object to an {@link objectArrayTrait}.
+   * @param stratumId The ID of the stratum in which to add the object.
+   * @param trait The name of the {@link objectArrayTrait} property.
+   * @returns The new object.
+   *
+   * The object will be placed at the end of the array (across all strata.
+   * This method takes `isRemoved` into account.
+   *
+   * This method can be used to add objects to `objectArrayTraits` which have `idProperty="index"` and support removals.
+   */
   pushObject<Key extends keyof ArrayElementTypes<T>>(
     stratumId: string,
     trait: Key
